Note.java implementation completed
Board.java implementation 70% completed

BoardRequest.java made to be constructed with a Board object passed as parameter.

BoardServer adjusted to account for command line args when program is run and to instantiate a board object.

Need to still:
Implement and include Board class
Implement and include Notes class
Create Server Responses
Use ObjectInputStream and ObjectOutputStream to transfer the Notes to server and client